pyproject.toml
[tool.poetry]
name = "blizzard-moment"
version = "0.1.0"
description = ""
authors = ["Aryan Kothari <87589047+thearyadev@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyautogui = "^0.9.54"
torch = {version = "^2.2.0", source="torch"}
numpy = "^1.26.4"
torchvision = {version = "^0.17.0", source="torch"}
pillow = "^10.2.0"
mypy = "^1.8.0"
types-pillow = "^10.2.0.20240213"
black = "^24.2.0"
imagehash = "^4.3.1"


[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cu121"
secondary = true

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
